+++ web/packages/api/src/toEthereum.ts
@@ -1,0 +1,4 @@
+
+export function send() {
+    return 42
+}
